Improve type signature for Parameter
